Added monitoring, extracting functionality and refactoring
Added a class that follows for now whether a new file has appeared in the
given folder that can be now provided as an argument. Moreover, whenever
a new .zip file appears, the parser will extract it in a local folder, then
extract the underlying .zip files contained in the initial .zip file, to
obtain the journal records themselves, in .txt files. After that, it parses
them and show the output.